Add an error output stream to the availability data source
This stream will make it possible for the data source to publish
events whenever it discovers that a resource is missing, for
instance in an async task running a WebSockets connection. This
can then be connected to a task which will fetch missing resources
from DA, without compromising the modularity of the data source
(since we have many data sources, but the logic to fetch missing
resources is the same for all of them).

In addition, it is just generally useful to have a way to surface
errors from background tasks, as we discovered when working on the
CAPE wallet GUI.