Set up SQL streams to be compatible with async DA fetching
The streams now contain bidirectional connections to an external
fetcher (which is left opaque and interacted with via message
passing). Whenever a stream encounters an item which is missing,
it sends a message to the fetcher requesting that the missing
object be retrieved. It then waits on the notification stream to
receive the requested object. Errors fetching are also distinguished
from missing objects and handled differently (by retrying).

This makes the subscription streams far more robust, and allows them
to yield simple objects, blocking until missing objects are
retrieved, rather than yielding objects wrapped in `QueryResult`.

New abstractions are introduced in the SQL data source so that we
can reuse most of the same code for the block and leaf streams, as
the retry and fetching logic is fairly complex.

Note that the changes apply only to the SQL data source. The file
system data source does not support out-of-order insertion of
objects anyways, so it will just fail if an object is missing. This
is the same behavior as before, and will be fixed in future work
(#16).